refactor(app): Separate everything in modules
Separate every file into their own modules
