+++ src/modules/props/Prop.ts
