+++ src/modules/props/PropDetail.ts
