+++ src/modules/props/PropDynamic.ts
