+++ src/modules/props/PropInteractable.ts
