+++ src/modules/props/PropStatic.ts
