+++ tailwind.config.ts
@@ -3,7 +3,6 @@ import type { Config } from 'tailwindcss'
 export default {
   content: [
     "./index.html",
-    "./src/ui/**/*.js",
   ],
   theme: {
     extend: {},
